+++ src/App.vue
@@ -45,6 +45,7 @@ watch(
 )
 
 onMounted(async () => {
+  localStorage.setItem('isLogin', userStore.user ? 'true' : 'false')
   await products.initProducts()
   await menuStore.getMenu()
 })