+++ CHANGELOG.md
@@ -14,6 +14,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change number of visibilities products in similar products.
 - Change style of button iva in the product view.
 
+## [1.6.3] - 08/07/2024
+
+### Fixed
+- Fix icon in loading firebase.
+
+## [1.6.2] - 08/07/2024
+
+### Fixed
+- Fix key `isLogin` localstorage in the project.
+
+### Changed
+- Change styles in modal login.
+- Change styles in `MpWhatsapp` component.
+
 ## [1.6.1] - 06/07/2024
 
 ### Fixed
@@ -163,6 +177,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add new version of the project. With products and services.
 
+[1.6.3]: https://github.com/CristhianDaza/megaprom/pull/18
+[1.6.2]: https://github.com/CristhianDaza/megaprom/pull/17
 [1.6.1]: https://github.com/CristhianDaza/megaprom/pull/16
 [1.6.0]: https://github.com/CristhianDaza/megaprom/pull/15
 [1.5.0]: https://github.com/CristhianDaza/megaprom/pull/14